--- projects/rust/fullstack-rustapp-template/packages/libs/svg_icons/src/svgs/mod.rs
@@ -1,4 +1,0 @@
-mod add_document;
-mod folder;
-mod grapes;
-mod panel_left;
